Turn maybe unused variable to static for compatibility with clang and gcc
Since clang warns on unused private variables and gcc does not
the maybe_unused attribute is ignored on gcc and therefore generated
a warning that is treated as an error. To makes our lives simpler,
let's turn this into a static variable which would produce almost
the same result for the context of this example.